requirements.txt
# Web框架
Django>=4.2.0,<5.0.0
whitenoise>=6.4.0

# 数据库
psycopg2-binary>=2.9.5
mysqlclient>=2.1.1

# 用户认证和权限
django-allauth>=0.54.0

# 图像处理
Pillow>=9.5.0

# HTTP请求处理
requests>=2.28.2

# 缓存和异步任务
django-redis>=5.2.0
redis>=4.5.4
celery>=5.2.7

# 开发工具
django-debug-toolbar>=4.0.0
ipython>=8.12.0

# 国际化
django-rosetta>=0.9.9

# API
djangorestframework>=3.14.0
markdown>=3.4.3
django-filter>=23.1

# 内容处理
bleach>=6.0.0 